requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
presidio-analyzer==2.2.358
presidio-anonymizer==2.2.358
spacy==3.7.2
redis==5.0.1
mysql-connector-python==8.2.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
phonenumbers==8.13.26


